# Conflicts:
#	.gitignore
#	CMake/Dependencies.cmake
#	CMake/makeExecutable.cmake
#	CMakeLists.txt
#	README.md
#	doc/Development.md
#	doc/Setup.md
#	doc/Usage.md
#	eval/README.md
#	eval/download_evaluation_data.py
#	eval/test-eval.py
#	extern/imgui/implot/implot.cpp
#	extern/imgui/implot/implot.h
#	extern/imgui/implot/implot_demo.cpp
#	extern/imgui/implot/implot_internal.h
#	extern/imgui/implot/implot_items.cpp
#	lib/vvv-glfw-app/include/vvvwindow/App.hpp
#	lib/vvv-glfw-app/include/vvvwindow/GLFWCameraController.hpp
#	lib/vvv-glfw-app/include/vvvwindow/GuiImgui.hpp
#	lib/vvv-glfw-app/include/vvvwindow/tf/TransferFunction1DWidget.hpp
#	lib/vvv-glfw-app/src/App.cpp
#	lib/vvv-glfw-app/src/GuiImgui.cpp
#	lib/vvv-glfw-app/src/entrypoint.cpp
#	lib/vvv-glfw-app/src/tf/TransferFunction1DWidget.cpp
#	lib/vvv-glfw-app/src/tf/TransferFunctionSegmentedVolumeWidget.cpp
#	lib/vvv/CMakeLists.txt
#	lib/vvv/include/vvv/core/Buffer.hpp
#	lib/vvv/include/vvv/core/Camera.hpp
#	lib/vvv/include/vvv/core/DefaultGpuContext.hpp
#	lib/vvv/include/vvv/core/GpuContext.hpp
#	lib/vvv/include/vvv/core/GuiInterface.hpp
#	lib/vvv/include/vvv/core/HeadlessRendering.hpp
#	lib/vvv/include/vvv/core/MultiBuffering.hpp
#	lib/vvv/include/vvv/core/Renderer.hpp
#	lib/vvv/include/vvv/core/Shader.hpp
#	lib/vvv/include/vvv/core/Synchronization.hpp
#	lib/vvv/include/vvv/core/Texture.hpp
#	lib/vvv/include/vvv/core/WindowingSystemIntegration.hpp
#	lib/vvv/include/vvv/core/WithGpuContext.hpp
#	lib/vvv/include/vvv/passes/PassBase.hpp
#	lib/vvv/include/vvv/passes/PassBlur.hpp
#	lib/vvv/include/vvv/passes/PassCompute.hpp
#	lib/vvv/include/vvv/passes/PassSimpleSsao.hpp
#	lib/vvv/include/vvv/passes/PassSsao.hpp
#	lib/vvv/include/vvv/passes/SinglePassGraphics.hpp
#	lib/vvv/include/vvv/reflection/UniformReflection.hpp
#	lib/vvv/include/vvv/util/Logger.hpp
#	lib/vvv/include/vvv/util/csv_utils.hpp
#	lib/vvv/include/vvv/util/random_gpu.hpp
#	lib/vvv/include/vvv/util/space_filling_curves.hpp
#	lib/vvv/include/vvv/vk/memory.hpp
#	lib/vvv/include/vvv/vk/version.hpp
#	lib/vvv/include/vvv/volren/Volume.hpp
#	lib/vvv/include/vvv/volren/tf/TransferFunction.hpp
#	lib/vvv/include/vvv/volren/tf/TransferFunction1D.hpp
#	lib/vvv/include/vvv/volren/tf/TransferFunction2D.hpp
#	lib/vvv/include/vvv/volren/tf/VectorTransferFunction.hpp
#	lib/vvv/src/core/Buffer.cpp
#	lib/vvv/src/core/Camera.cpp
#	lib/vvv/src/core/DefaultGpuContext.cpp
#	lib/vvv/src/core/GpuContext.cpp
#	lib/vvv/src/core/GuiElementList.cpp
#	lib/vvv/src/core/GuiInterface.cpp
#	lib/vvv/src/core/HeadlessRendering.cpp
#	lib/vvv/src/core/MultiBuffering.cpp
#	lib/vvv/src/core/Shader.cpp
#	lib/vvv/src/core/Texture.cpp
#	lib/vvv/src/headless_entrypoint.cpp
#	lib/vvv/src/passes/PassBlur.cpp
#	lib/vvv/src/passes/PassSimpleSsao.cpp
#	lib/vvv/src/passes/PassSsao.cpp
#	lib/vvv/src/reader/volcanite_raw.cpp
#	lib/vvv/src/reader/vti.cpp
#	lib/vvv/src/util/random_gpu.cpp
#	lib/vvv/src/util/util.cpp
#	lib/vvv/src/vk/format_utils.cpp
#	lib/vvv/src/vk/memory.cpp
#	lib/vvv/src/vk/queue.cpp
#	lib/vvv/src/volren/tf/VectorTransferFunction.cpp
#	python/convert.py
#	python/volcanite/VolcaniteEval.md
#	python/volcanite/pyproject.toml
#	python/volcanite/src/volcanite/converter.py
#	python/volcanite/src/volcanite/converter_chunked.py
#	python/volcanite/src/volcanite/volcaniteeval.py
#	test/bitvector_test.cpp
#	test/csgv_gpu_test.cpp
#	test/csgv_test.cpp
#	test/volcanite_render_test.cpp
#	test/wavelet_matrix_test.cpp
#	volcanite/CMakeLists.txt
#	volcanite/data/shader/cpp_glsl_include/csgv_constants.incl
#	volcanite/data/shader/volcanite/bit_vector.glsl
#	volcanite/data/shader/volcanite/compression/decoder/HuffmanWMDecoder.glsl
#	volcanite/data/shader/volcanite/renderer/csgv_assign.comp
#	volcanite/data/shader/volcanite/renderer/csgv_bindings.glsl
#	volcanite/data/shader/volcanite/renderer/csgv_decompress.comp
#	volcanite/data/shader/volcanite/renderer/csgv_denoise_resolve.comp
#	volcanite/data/shader/volcanite/renderer/csgv_provision.comp
#	volcanite/data/shader/volcanite/renderer/csgv_renderer.comp
#	volcanite/data/shader/volcanite/renderer/csgv_request.comp
#	volcanite/data/shader/volcanite/renderer/dummy_envmap.glsl
#	volcanite/data/shader/volcanite/renderer/framebuffer.glsl
#	volcanite/include/volcanite/CSGVPathUtils.hpp
#	volcanite/include/volcanite/VolcaniteArgs.hpp
#	volcanite/include/volcanite/compression/CSGVChunkMerger.hpp
#	volcanite/include/volcanite/compression/CSGVDatabase.hpp
#	volcanite/include/volcanite/compression/CompSegVolHandler.hpp
#	volcanite/include/volcanite/compression/CompressedSegmentationVolume.hpp
#	volcanite/include/volcanite/compression/VolumeCompressionBase.hpp
#	volcanite/include/volcanite/compression/encoder/CSGVBrickEncoder.hpp
#	volcanite/include/volcanite/compression/encoder/CSGVSerialBrickEncoder.hpp
#	volcanite/include/volcanite/compression/encoder/RangeANSEncoder.hpp
#	volcanite/include/volcanite/compression/encoder/WaveletMatrixEncoder.hpp
#	volcanite/include/volcanite/compression/memory_mapping.hpp
#	volcanite/include/volcanite/compression/pack_nibble.hpp
#	volcanite/include/volcanite/compression/pack_rans.hpp
#	volcanite/include/volcanite/compression/pack_wavelet_matrix.hpp
#	volcanite/include/volcanite/compression/wavelet_tree/BitVector.hpp
#	volcanite/include/volcanite/compression/wavelet_tree/HuffmanWaveletMatrix.hpp
#	volcanite/include/volcanite/compression/wavelet_tree/WaveletMatrix.hpp
#	volcanite/include/volcanite/compression/wavelet_tree/WaveletMatrixBase.hpp
#	volcanite/include/volcanite/compression/wavelet_tree/prefix_counting.hpp
#	volcanite/include/volcanite/eval/CSGVBenchmarkPass.hpp
#	volcanite/include/volcanite/eval/EvaluationLogExport.hpp
#	volcanite/include/volcanite/renderer/CompressedSegmentationVolumeBrickViewer.hpp
#	volcanite/include/volcanite/renderer/CompressedSegmentationVolumeRenderer.hpp
#	volcanite/include/volcanite/renderer/HelloVolumeRenderer.hpp
#	volcanite/include/volcanite/renderer/PassCompSegVolRender.hpp
#	volcanite/include/volcanite/util/args_and_csgv_provider.hpp
#	volcanite/include/volcanite/util/segmentation_volume_synthesis.hpp
#	volcanite/package_assets/LICENSE_THIRD_PARTY.txt
#	volcanite/src/bin/csgv_benchmark.cpp
#	volcanite/src/bin/csgv_brick_viewer.cpp
#	volcanite/src/bin/volcanite.cpp
#	volcanite/src/renderer/CompressedSegmentationVolumeBrickViewer.cpp
#	volcanite/src/renderer/CompressedSegmentationVolumeRenderer.cpp
#	volcanite/src/renderer/PassCompSegVolRender.cpp
#	volcanite/src/util/segmentation_volume_synthesis.cpp

+++ lib/vvv/src/core/GuiInterface.cpp
@@ -79,7 +79,7 @@ void GuiInterface::GuiTFSegmentedVolumeEntry::updateVectorColormap(int material)
         break;
     case GuiInterface::GuiTFSegmentedVolumeEntry::SVTFImport: {
         mat.tf->m_interpolationColorSpace = VectorTransferFunction::RGB;
-        int targetSizeControlPointsRgb = d.color.size() * 4;
+        size_t targetSizeControlPointsRgb = d.color.size() * 4;
         if (mat.tf->m_controlPointsRgb.size() != targetSizeControlPointsRgb) {
             mat.tf->m_controlPointsRgb.resize(targetSizeControlPointsRgb);
         }
